% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_plots}
\alias{create_plots}
\title{create_plots.R}
\usage{
create_plots(plotData)
}
\arguments{
\item{plotData}{data.frame with columns labeled (exactly) Chr, BP, cum_avg_lfdr, Z_Dataset1, Z_Dataset2}
}
\value{
A list with the elements:
\item{manPlot}{A ggplot object of the Manhattan plot of the lfdr-values at which each SNP is significant}
\item{zPlot}{A ggplot object of a scatterplot comparing the z-scores for the first two datasets}
}
\description{
Creates lfdr-value Manhttan plots and z-score plots to visualize composite null hypothesis testing results
}
\examples{
set.seed(0)
plotData <- data.frame(Chr = rep(1:22, each=100), BP=runif(n=2200, min=1, max=10^8),
cum_avg_lfdr=runif(n=2200, min=0, max=22),
Z_Dataset1=rnorm(n=2200), Z_Dataset2=rnorm(n=2200))
create_plots(plotData = plotData)

}
