% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMaxMeans.R
\name{find_max_means}
\alias{find_max_means}
\title{find_max_means.R}
\usage{
find_max_means(muInfo)
}
\arguments{
\item{muInfo}{A list with 2^K elements, where each element is a matrix with K rows and Mb columns.}
}
\value{
A K*1 vector of the maximum means for each dimension under the null.
}
\description{
Find maximum means for each dimension in null settings.
}
\examples{
initMuList <- list(matrix(data=0, nrow=2, ncol=1), matrix(data=c(0, 3, 0, 6), nrow=2),
matrix(data=c(3, 0, 6, 0), nrow=2), matrix(data=c(8, 8), nrow=2))
find_max_means(initMuList)
}
