% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_init_lists.R
\name{generate_init_lists}
\alias{generate_init_lists}
\title{generate_init_lists.R}
\usage{
generate_init_lists(K)
}
\arguments{
\item{K}{Scalar, number of datasets in the analysis.}
}
\value{
A list with the elements:
\item{initMuList}{List of 2^K elements where each element is a matrix with K rows and number of columns equal to the number of possible mean vectors for that binary representation}
\item{initPiList}{List of 2^K elements where each element is a vector with number of elements equal to the number of possible mean vectors for that binary representation.}
}
\description{
Generates initMuList and initPiList parameters for symm_fit_ind_EM() function.
}
\examples{
K <- 2
generate_init_lists(K)
}
