% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_csmgmm_data.R
\name{prepare_csmgmm_data}
\alias{prepare_csmgmm_data}
\title{prepare_csmgmm_data.R}
\usage{
prepare_csmgmm_data(datasets, dataset_names = NULL, z_cap = 8.1)
}
\arguments{
\item{datasets}{list holding K datasets to use for replication analysis}

\item{dataset_names}{optional vector of names for each dataset; otherwise function uses Dataset1,.., DatasetK as default.}

\item{z_cap}{optional number to cap the most extreme z-scores, we recommend setting at 8.1 as R is not very accurate for very large z-scores}
}
\value{
A list with the elements:
\item{extra_dat}{A J*K matrix with J matching SNPs across K datasets, holds relevant SNP information.}
\item{clean_dat}{J*K matrix of test statistics, holds J Z-scores for each dataset.}
\item{n_snps}{A scalar returning the number of matching SNPs across all K datasets.}
}
\description{
Prepares J*K matrix of summary statistics for GWAS replication studies where J is the number of SNPs and K is the number of cohorts.
}
\examples{
library(dplyr)
set.seed(0)
J <- 1000
alleles <- c("A", "C", "T", "G")
dat1 <- data.frame(zstat = rnorm(J)) \%>\% mutate(se = runif(n=J, min=0.0005, max=0.0015)) \%>\%
mutate(beta = zstat * se, chr=1, pos=1:J, a1=sample(x=alleles, J, replace=TRUE),
a2=sample(x=alleles, J, replace=TRUE))
dat2 <- data.frame(zstat = rnorm(J)) \%>\% mutate(se = runif(n=J, min=0.0005, max=0.0015)) \%>\%
mutate(beta = zstat * se, chr=1, pos=1:J, a1=sample(x=alleles, J, replace=TRUE),
a2=sample(x=alleles, J, replace=TRUE))
dat3 <- data.frame(zstat = rnorm(J)) \%>\% mutate(se = runif(n=J, min=0.0005, max=0.0015)) \%>\%
mutate(beta = zstat * se, chr=1, pos=1:J, a1=sample(x=alleles, J, replace=TRUE),
a2=sample(x=alleles, J, replace=TRUE))
datasets = list(dat1, dat2, dat3)
prepare_csmgmm_data(datasets = datasets)

}
