% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_lfdr_results.R
\name{process_lfdr_results}
\alias{process_lfdr_results}
\title{process_lfdr_results.R}
\usage{
process_lfdr_results(orig_data, lfdrResults, fdr_threshold = 0.1)
}
\arguments{
\item{orig_data}{data.frame with J rows, one for each composite null hypothesis being tested, holds data such as position, z-statistics, etc.}

\item{lfdrResults}{J*1 vector of all lfdr-values, the jth value corresponds to the jth row of orig_data.}

\item{fdr_threshold}{Scalar between 0 and 1 determining the percentage of false discoveries, default set to 0.1.}
}
\value{
A list with elements:
\item{all_snps_info}{A data.frame of all SNPs, sorted by the lfdr-value at which they are significant (cum_avg_lfdr column)}
\item{sig_snps_info}{A data.frame of the significant SNPs, sorted by the lfdr-value at which they are significant (cum_avg_lfdr column)}
\item{n_significant}{A scalar returning the number of significant SNPs.}
}
\description{
Sorts lfdrResults and calculates cumulative averages to return significant SNPs.
}
\examples{
set.seed(0)
orig_data <- data.frame(
variant_id = sample(c("A","B","C","D"), 10^4, replace = TRUE),
Chr = sample(1:22, 10^4, replace = TRUE),
BP = sample(1e6:2e6, 10^4),
Z_Dataset1  = rnorm(10^4),
Z_Dataset2    = rnorm(10^4)
)
lfdrResults <- runif(10^4)
process_lfdr_results(orig_data = orig_data, lfdrResults = lfdrResults, fdr_threshold = 0.05)
}
