% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gwas.R
\name{read_gwas_set}
\alias{read_gwas_set}
\title{Read 2-5 GWAS Catalog summary-statistics files into the current environment.}
\usage{
read_gwas_set(urls, ...)
}
\arguments{
\item{urls}{Character vector of 2-5 https URLs to .tsv.gz files.
If the vector has names, those names are used in the returned
list; otherwise entries are named gwas1, gwas2, ...}

\item{...}{Forwarded to read_gwas_one (cache_dir, overwrite, max_tries,
fread_args).}
}
\value{
Named list of data.tables, one per input URL.
}
\description{
Read 2-5 GWAS Catalog summary-statistics files into the current environment.
}
\examples{
\donttest{
# Ensure internet is available before running the example
if (curl::has_internet()) {
  try({
    dat <- read_gwas_set(c(
      paste0("http://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/",
      "GCST005001-GCST006000/GCST005527/tsoi_2012_23143594_pso_efo0000676_1_ichip.sumstats.tsv.gz"),
      paste0("http://ftp.ebi.ac.uk/pub/databases/gwas/summary_statistics/",
      "GCST003001-GCST004000/GCST003044/CD_trans_ethnic_association_summ_stats_b37.txt.gz")
    ))
    gwas1 <- dat$gwas1
    gwas2 <- dat$gwas2
    head(gwas1)
  })
}
}
}
